# Introduce StaticRouter and SmartRouter