--- a/deno_dist/router/smart-router/index.ts
+++ b/deno_dist/router/smart-router/index.ts
@@ -1,0 +1,1 @@
+export { SmartRouter } from './router.ts'