--- a/deno_dist/router/static-router/index.ts
+++ b/deno_dist/router/static-router/index.ts
@@ -1,0 +1,1 @@
+export { StaticRouter } from './router.ts'