--- a/src/router.ts
+++ b/src/router.ts
@@ -1,5 +1,6 @@
 export const METHOD_NAME_ALL = 'ALL' as const
 export const METHOD_NAME_ALL_LOWERCASE = 'all' as const
+export const METHODS = ['get', 'post', 'put', 'delete', 'head', 'options', 'patch'] as const
 
 export interface Router<T> {
   add(method: string, path: string, handler: T): void