--- a/src/router/smart-router/index.ts
+++ b/src/router/smart-router/index.ts
@@ -1,0 +1,1 @@
+export { SmartRouter } from './router'