--- a/src/router/static-router/index.ts
+++ b/src/router/static-router/index.ts
@@ -1,0 +1,1 @@
+export { StaticRouter } from './router'