### Motivation

I think it would be difficult for users to take advantage of the "hono has two routers and the optional RegExpRouter is faster."
This PR changes this to "By default, the best router is selected based on the routing of the application."


### StaticRouter

StaticRouter is a new alternative. Works fast when there is no path to receive parameters.
I think that this is not a router for benchmarking purposes, and that it could be used in the real world as follows

```ts
app.use('*', logger(), cors())
app.use(
  '/graphql',
  graphqlServer({
    schema,
    rootResolver,
  })
)
```

However, when benchmarked, there did not appear to be a significant difference with RegExpRouter in nodejs.
If it is a bun, there may be a little difference, but there may not be.

I think this StaticRouter is a good example of the use of SmartRouter, but may not be necessary for hono.


### SmartRouter

This is a router that finds the best one for routing from several possible routers and invoke it when `match` is called.


### Side effects of merging this

* All routers are bundled by default, which increases the bundle size.
* Slightly longer spinup time.
* Makes debugging routers more difficult.

### Benchmark

#### with path parameters

https://gist.github.com/usualoma/6d8db87c9f40614073b0517dc3e8cee6

```
hono - smart x 267,322 ops/sec ±5.61% (77 runs sampled)
hono - smart - trie-router x 230,474 ops/sec ±5.72% (81 runs sampled)
hono - smart - regexp-router x 267,080 ops/sec ±5.65% (68 runs sampled)
hono - trie-router x 223,486 ops/sec ±5.05% (84 runs sampled)
hono - regexp-router x 262,344 ops/sec ±6.61% (77 runs sampled)
Fastest is hono - smart,hono - smart - regexp-router,hono - regexp-router
```

#### without path parameters

https://gist.github.com/usualoma/0b86e02ee9f3a16dd8e5af77f479624c

```
hono - smart - trie-router x 276,017 ops/sec ±6.15% (77 runs sampled)
hono - smart - regexp-router x 298,234 ops/sec ±6.67% (75 runs sampled)
hono - smart - static-router x 296,272 ops/sec ±6.79% (63 runs sampled)
hono - trie-router x 267,239 ops/sec ±6.06% (81 runs sampled)
hono - regexp-router x 303,253 ops/sec ±5.80% (82 runs sampled)
Fastest is hono - regexp-router,hono - smart - regexp-router,hono - smart - static-router
```

#### bun - RegExpRouter

```ts
import { Hono } from './src/index'
import { RegExpRouter } from './src/router/reg-exp-router'

const app = new Hono({ router: new RegExpRouter() })

app.get('/posts', (c) => c.text('foo'))

export default app
```

```
$ ./bombardier -c 200 -d 20s http://10.1.0.139:3000/posts
Bombarding http://10.1.0.139:3000/posts for 20s using 200 connection(s)
[==========================================================================] 20s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     72705.89    1742.91   75809.44
  Latency        2.75ms    86.69us    10.99ms
  HTTP codes:
    1xx - 0, 2xx - 1454123, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    10.40MB/s
```


#### bun - SmartRouter - StaticRouter

```ts
import { Hono } from './src/index'
import { SmartRouter } from './src/router/smart-router'
import { StaticRouter } from './src/router/static-router'

const app = new Hono({
  router:
    new SmartRouter({
      routers: [new StaticRouter()],
    })
})

app.get('/posts', (c) => c.text('foo'))

export default app
```

```
$ ./bombardier -c 200 -d 20s http://10.1.0.139:3000/posts
Bombarding http://10.1.0.139:3000/posts for 20s using 200 connection(s)
[==========================================================================] 20s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     74691.57    1911.30   80312.78
  Latency        2.68ms    97.25us    10.77ms
  HTTP codes:
    1xx - 0, 2xx - 1493610, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    10.68MB/s
```