@yusukebe 2022-09-06
Wow, crazy!

I'm going to look at the implementation and consider it well.

> All routers are bundled by default, which increases the bundle size.

This is what I am concerned about.

Again!