@usualoma 2022-09-06
There are no more commits to be added.

I think it's an interesting and useful idea, but I'm not sure I really want to merge this.