@yusukebe 2022-09-06
Hi @usualoma !

This idea is ultra great! I like it! But as you are thinking, we have to consider well whether to merge or not. So, I don't make conclusions right away.

I investigated "bundle size". `wrangler` command will bundle the scripts and shows the "upload size". Below are the sizes of "Hello World" program on each framework for Cloudflare Workers.

```
Hono v2.1.4
Total Upload: 27.04 KiB / gzip: 6.09 KiB

This PR / with SmartRouter
Total Upload: 50.35 KiB / gzip: 10.32 KiB

Sunder
Total Upload: 47.21 KiB / gzip: 11.08 KiB

Worktop
Total Upload: 7.02 KiB / gzip: 2.70 KiB
```

This is just "information", and I would still like to think about it.