@yusukebe 2022-09-07
In actuality, there is a way to reduce the bundle size.

Currently, we build the project as "CommonJS". If we make it ESM, the size will be smaller. I've tried to change the compiler option on `tscofing.json` and build it as ESM. Then, the bundle size built by `wrangler` has been smaller!

```
Total Upload: 36.10 KiB / gzip: 8.64 KiB
```

I will create another Issue on the matter of migrating from CommonJS to ESM. But, it's good news.