@usualoma 2022-09-07
@yusukebe Good news!

```
Hono this PR / with SmartRouter
Total Upload: 36.10 KiB / gzip: 8.64 KiB
```

Hmmm, but RegExpRouter is so huge.