@yusukebe 2022-09-10
Hi @usualoma !

I want to merge this Static Router and Smart Router! I love this feature.

> I think it would be difficult for users to take advantage of the "hono has two routers and the optional RegExpRouter is faster." This PR changes this to "By default, the best router is selected based on the routing of the application."

You are right about this. Users have to use the fastest router without having to do anything else.

I love it because this is **really "smart"**. It will be more difficult to debug, but it will be more "fun" with this way.

I think the issue of the bundle size could be improved a little by migrating ComonJS to ESM. I don't know if it will work, but I would like to try it.

How about we go with this!