@usualoma 2022-09-11
@yusukebe 
Sorry for the delay in answering.

Thanks! I agree with merging.

I've been thinking the past few days that if we assume SmartRouter, we could create a lightweight version of RegExpRouter with some features removed from RegExpRouter. (I haven't gotten around to starting it, though.)

I think we can represent RegExpRouter with less code by deciding not to support it in the lightweight version of RegExpRouter, since there are rarely routings that have different variable names in the same place, such as the following.

```
app.get('/path/:var1/path')
app.get('/path/:var2/path')
```

I would like to try such a refactoring along with migrating ComonJS to ESM.

Shall we fix this branch so that the test is green and merge it into main?