@yusukebe 2022-09-11
Hi @usualoma !

> I think we can represent RegExpRouter with less code by deciding not to support it in the lightweight version of RegExpRouter, since there are rarely routings that have different variable names in the same place, such as the following.

This is a great idea! Just fall back to TrieRouter If it hits such rare cases.

> Shall we fix this branch so that the test is green and merge it into main?

I've created the `develop` branch <https://github.com/honojs/hono/tree/develop>

Let's work on the `develop`" branch for the new features such as migration to SmartRouter and ESM, will be included in v2.2.0 new version.  So, fix this branch and merge it into the `develop` branch.